+++ docs/source/changelog.rst
@@ -2,6 +2,14 @@
 CHANGELOG
 =========
 
+XXXX.XX.XX
+----------
+
+**Improvements:**
+
+- Include basic certificates in docker image
+
+
 2023.4.3       (2023-04-13)
 ---------------------------