+++ docs/source/changelog.rst
@@ -10,6 +10,11 @@ CHANGELOG
 - Improve admin to define polygons patterns
 - Filter frontend with non empty views
 - Default view is now the first ordered for an user (authenticated or not)
+- Disable autocomplete / autofill on PostGIS source form in admin
+
+**Bugfix**
+
+- Allow PostGIS source form edition in admin without retype password
 
 
 2023.4.3       (2023-04-13)